==> IMDB movie Ratings/1 Cleaning and merging Data part -1 .ipynb <==
import pandas as pd

# Loading Files of IMDB movie dataset

#Loading files through for loop 

#file_names = ["C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\1name_data.tsv",
             # "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\2title_akas.tsv",
             # "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\3title_basics.tsv",
             # "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\4title_crew.tsv",
              #"C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\5title_episode.tsv",
             # "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\6title_principle.tsv",
             # "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\7title_ratings.tsv"]         

#fileData = pd.read_csv(file_names[0], sep='\t')
#fileData.head(2)

#for fileName in file_names:
    #fileData = pd.read_csv(fileName, sep='\t')
    #fd = fileData.head(2)
    #print(fd)
    #break;
    #print(fileData.head(2))
    #fileData.head()

# Loading files individaully
name_data = "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\1name_data.tsv"
title_akas = "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\2title_akas.tsv"
title_basics = "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\3title_basics.tsv"
title_crew = "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\4title_crew.tsv"
title_episode = "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\5title_episode.tsv"
tilte_principle = "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\6tiltle_principle.tsv"
title_ratings = "C:\\Users\\vishd\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\7title_ratings.tsv"

#loading title_akas files 

title_akas = pd.read_csv(title_akas, sep='\t')                                         
title_akas.head()


# Dropping duplicate

title_akas_dropped = title_akas.drop_duplicates('titleId',keep=False)                    
title_akas_dropped.head()

# Loading title basics files 

title_basics = pd.read_csv(title_basics, sep='\t')
title_basics.head(2)
#title_basics.shape

# Merging - 1 

# merging both above files (title_akas and title_basics) the shape of the merged df is 7794885, 17. 

join1 = title_akas_dropped.merge(title_basics, left_on='titleId', right_on='tconst', how='outer')
join1.shape                                                                       

# Dropping column with duplicate and no values
# Checking unique values in columns and dropping them 
# Dropping columns with no or less values

#join1['isOriginalTitle'].nunique(dropna = True)                 # Checking unique values in columns and dropping them) 
#[languages, types, attributes, tconst, isAdult, primaryTitle, originalTitle, 'endyear(with 96)',runtimeMinutes(677), 'ordering(1)' ]  

join1 =join1.drop(['language','types', 'attributes', 'tconst', 
            'isAdult', 'primaryTitle', 'originalTitle',
            'isOriginalTitle', 'endYear', 'runtimeMinutes','ordering'], axis=1)
                                                                
join1.head()
join1.shape

# Loading title_crew files and merging with above Join1 df 

title_crew = pd.read_csv(title_crew, sep='\t')
title_crew.head(2)
#title_crew.shape

# Merging 2

# merging title_crew with previous Join1 df

join2 = join1.merge(title_crew, left_on='titleId', right_on='tconst', how='outer')        
join2.head()

# Dropping Duplicate

join2 =join2.drop(['tconst'], axis=1)
#join2.head()
join2.shape

# Loading and reading title_episode

title_episode = pd.read_csv(title_episode, sep='\t')
title_episode.head(2)
#title_episode.shape

# Merging 3

# merging title_episode with previous Join2 df

Join3 = join2.merge(title_episode, left_on='titleId', right_on='tconst', how='outer') # merging title_episode with previous df
Join3.head()

# dropping duplicate columns 

Join3 =Join3.drop(['tconst'], axis=1)
Join3.head()
#Join3.shape


#reading and Loading title_ratings 

title_ratings = pd.read_csv(title_ratings, sep='\t')
title_ratings.head(2)

# Merging - 4

# merging title_rating with Previous Join3 df

Join4 = Join3.merge(title_ratings, left_on='titleId', right_on='tconst', how='outer') 
Join4.head()


# dropping duplicate columns 

Join4 =Join4.drop(['tconst'], axis=1)
Join4.head()
#print(Join4.shape)

# Loading and reading name_data files 

name_data = pd.read_csv(name_data, sep='\t')
#name_data.head(2)  


# rename column's title name 

name_data = name_data.rename(columns={'primaryName': 'primaryName_Director', 'birthYear':'birthYear_Director', 
                          'death_Year':'death_year_director', 'primaryProfession':'primaryProfession_director',
                          'knownForTitles':'Dir_knownForTitles'})                 # renaming columns before merge
name_data.head()

# Merging 5

#merging renamed name_data df with prvious df join 4

Final_DF = Join4.merge(name_data.head(), left_on='directors', right_on='nconst', how='outer') 
                                                                            
Final_DF.head()


# dropping duplicate col from merged df

Final_DF = Final_DF.drop(['nconst'], axis=1)                                     
Final_DF

Final_DF.shape                                                 

# Saving Final_DF as csv file 

Final_DF.to_csv('Final_DF1.csv')
