==> IMDB movie Ratings/2 IMDB data wrangling and Cleaning Part 2.ipynb <==
import pandas as pd

# Loading Merged IMDB dataset

Final_IMDB_data = "C:\\Users\\vishd\\Desktop\\reetu\\Springboard course\\Project 1\\Final_DF1.csv"
Final_IMDB_data = pd.read_csv(Final_IMDB_data)#, sep='\t')                                                 # loading first files
Final_IMDB_data.head()


# Checking shape of the DataFrame

# Final_IMDB_data.shape

# Removing Duplicates

IMDB_data_NoDup = Final_IMDB_data.drop_duplicates(subset=None, keep='first')                              
IMDB_data_NoDup.shape

# dropping duplicated column of AveRating and numvotes

IMDB_data_NoDup = IMDB_data_NoDup.drop(columns=['averageRating_y', 'numVotes_y'])

# Rename col AveRating and numvotes
IMDB_data_NoDup = IMDB_data_NoDup.rename(columns={'Unnamed: 0': 'Index'})                        



IMDB_data_NoDup = IMDB_data_NoDup.rename(columns={'averageRating_x': 'AveRating', 'numVotes_x': 'numVotes'})
IMDB_data_NoDup.head(2)

# Setting Index and renaming Unnamed columns

IMDB_data_NoDup =IMDB_data_NoDup.set_index('Unnamed: 0', drop=True, append=False, inplace=False, verify_integrity=False)
IMDB_data_NoDup.head()

IMDB_data_NoDup = IMDB_data_NoDup.rename(columns={'Unnamed: 0': 'Index'})                        
IMDB_data_NoDup.head()
IMDB_data_NoDup.shape


# Handling for null values 

# cheking for columns with null values 

IMDB_data_NoDup.columns[IMDB_data_NoDup.isnull().any()]


#Checking sum of the null values 

for column in IMDB_data_NoDup:                                                             
    IMDB_data_missing = IMDB_data_NoDup[column].isnull().sum()
    #print(IMDB_data_missing)
    #print(column, IMDB_data_NoDup[column].isnull().sum())

Calculating the % of missing vlaues 

percent_NA_cols = IMDB_data_missing / len(IMDB_data_NoDup)
percent_NA_cols

import numpy as np
IMDB_data_NoDup.dropna(thresh=np.int((100-percent_NA_cols)*(len(IMDB_data_NoDup.columns)/100)),inplace=True)
print(IMDB_data_NoDup.shape)
IMDB_data_NoDup.head(5)
#for col in IMDB_data_NoDup.columns: 
   # if sum(IMDB_data_NoDup[col].isnull())/float(len(IMDB_data_NoDup)) > 0.8:
       # del IMDB_data_NoDup[col] 
#IMDB_data_NoDup.head()

# Rename unnammed index to index 

IMDB_data_NoDup.index.names = ['Index']
IMDB_data_NoDup.head(2)

# Checking datatypes of each columns 

IMDB_data_NoDup.dtypes                                                            

# Checking Unique values

IMDB_data_NoDup.nunique()                                                            #printing Unique value

#IMDB_data_NoDup = IMDB_data_NoDup.dropna()
#IMDB_data_NoDup.shape

# Filling missing values 

IMDB_data_NoDup.fillna(method='bfill', inplace=True)                                   #Replacing missing values
IMDB_data_NoDup.shape

IMDB_data_NoDup['title'].unique()

IMDB_data_NoDup['startYear'].unique()

IMDB_data_NoDup['startYear'] = pd.to_datetime(IMDB_data_NoDup['startYear'], format='%y', errors='coerce')
IMDB_data_NoDup.head(5)
#IMDB_data_NoDup['startYear'] = pd.to_datetime(IMDB_data_NoDup[['startYear']]).year
#IMDB_data_NoDup.head()
#print(IMDB_data_NoDup['startYear'].unique())


IMDB_data_NoDup.loc[IMDB_data_NoDup['startYear'] == '\\N' , 'startYear'] = '0000'
IMDB_data_NoDup.loc[IMDB_data_NoDup['startYear'] == 'nan' , 'startYear'] = '0000'

IMDB_data_NoDup[IMDB_data_NoDup['startYear'] == '0000'].head(5) 



IMDB_data_NoDup.dtypes


Curremt_year = 2021
IMDB_data_NoDup.loc[:, "Age_of_movie"] = IMDB_data_NoDup["startYear"].apply(lambda x: 2021 - x)
IMDB_data_NoDup.head()

IMDB_data_NoDup.to_csv('IMDB_data_NoDup.csv')
