==> IMDB movie Ratings/4 Data handling for  Preprocessing and  Scaling .ipynb <==
# Importing Libraries

import pandas as pd
import math
import numpy as np
from scipy.stats import norm
from scipy import stats

from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import classification_report
from sklearn.model_selection import train_test_split
from sklearn.metrics import roc_auc_score
from sklearn.model_selection import GridSearchCV
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

#from google.colab import drive
#drive.mount('/IMDB_data_NoDup.csv')

# Importing and reading csv movie dataset 
The size of the file is too big 350039x16. So will continue only with the 5000 data. 

IMDB_data_NoDup = "C:\\Users\\vishd\\Desktop\\reetu\\Springboard course\\Data wrangling IMDB dataset\\IMDB_data_NoDup.csv"
#IMDB_data_NoDup.head()
#IMDB_data_NoDup = "/content/IMDB_data_NoDup.csv"
#IMDB_data_NoDup = pd.read_csv(IMDB_data_NoDup, sep='\t')
#IMDB_data_NoDup.shape
IMDB_data_NoDup = pd.read_csv(IMDB_data_NoDup, sep='\t', nrows = 5000)#, delimiter=',', header=None, skiprows=1)#, names=['LABEL','F1','F2','F3','F4','F5','X','Y','Z','C1','C2'])                                                # loading first files
IMDB_data_NoDup.head()


# Droping one index col. 
Here we have two columns with index #. Need to drop index. 

IMDB_data_NoDup = IMDB_data_NoDup.drop(['Index'], axis = 1)                                             #dropping index column


# Basic Statistics 
Below is the basis statistics for our data set

IMDB_data_NoDup.describe()

# Converting years  to numeric columns

#converting year col ot numeric type

IMDB_data_NoDup['startYear'] = pd.to_numeric(IMDB_data_NoDup['startYear'], errors='coerce')
IMDB_data_NoDup['startYear'] = IMDB_data_NoDup["startYear"].convert_dtypes() 

IMDB_data_NoDup['startYear'].unique()

#Converting death year col to numeric

IMDB_data_NoDup['deathYear'] = pd.to_numeric(IMDB_data_NoDup["deathYear"], errors='coerce')
IMDB_data_NoDup['deathYear'] = IMDB_data_NoDup["deathYear" ].convert_dtypes() 

# converting birthyear_Director col ot Numeric

IMDB_data_NoDup['birthYear_Director'] = pd.to_numeric(IMDB_data_NoDup["birthYear_Director"], errors='coerce')
IMDB_data_NoDup['birthYear_Director'] = IMDB_data_NoDup["birthYear_Director"].convert_dtypes() 


IMDB_data_NoDup.head()

# We can see the type of startYear is int

IMDB_data_NoDup['startYear'].dtype

# Calculating the age of the movie

Cuurent_year = 2021
IMDB_data_NoDup['Age_of_movie'] = Cuurent_year - IMDB_data_NoDup['startYear'] 

IMDB_data_NoDup.head()

IMDB_data_NoDup['Age_of_movie'].unique() 

IMDB_data_NoDup['birthYear_Director'].unique()

# Calculating the age of the directors

IMDB_data_NoDup['Director_age'] = IMDB_data_NoDup['deathYear']  - IMDB_data_NoDup['birthYear_Director'] 
IMDB_data_NoDup.head(2)

# Calculating std, min, max, mean of the rating columns 

std_Of_AvRating = np.std(IMDB_data_NoDup['AveRating'])
mean_mu = np.mean(IMDB_data_NoDup['AveRating'])
min_Rating = np.min(IMDB_data_NoDup['AveRating'])
max_Rating = np.max(IMDB_data_NoDup['AveRating'])
lowerbound, upperbound = stats.norm.interval(0.95, loc=mean_mu, scale= std_Of_AvRating)

print('std_Of_AvRating:',std_Of_AvRating, 'mean_mu of rating:', mean_mu)
print('min_Of_AvRating:',min_Rating, 'max_Of_AvRating:',min_Rating)
print('lowerbound:', lowerbound, 'upperbound', upperbound)

#IMDB_data_NoDup['startYear'] = IMDB_data_NoDup['startYear'].astype(str).replace('\.0', '', regex=True)
#IMDB_data_NoDup['startYear'] = IMDB_data_NoDup['startYear'].astype(str).replace('\N', '', regex=True)

# Checking and filling NaN values 

#Checking NaN values
IMDB_data_NoDup['Age_of_movie'].isna().sum()


#Converting Age_of_movie to float type
IMDB_data_NoDup['Age_of_movie'] = IMDB_data_NoDup['Age_of_movie'].astype('float')

#Replacing all NaN value with mean of the Age_of_movie column 
IMDB_data_NoDup['Age_of_movie'].fillna((IMDB_data_NoDup['Age_of_movie'].mean()), inplace=True)

IMDB_data_NoDup['Age_of_movie'].value_counts()

# dividing Age_of_movie col to bins(decade) of the movies


conditions = [
    # (IMDB_data_NoDup['Age_of_movie'].ndim == 0),
     (IMDB_data_NoDup['Age_of_movie'] <= 1),
     (IMDB_data_NoDup['Age_of_movie'] > 1) & (IMDB_data_NoDup['Age_of_movie'] <= 10), 
     (IMDB_data_NoDup['Age_of_movie'] > 10) & (IMDB_data_NoDup[ 'Age_of_movie'] <= 25),
     (IMDB_data_NoDup['Age_of_movie'] > 25) & (IMDB_data_NoDup['Age_of_movie'] <= 50),
     (IMDB_data_NoDup['Age_of_movie'] > 50) & (IMDB_data_NoDup['Age_of_movie'] <= 75),
     (IMDB_data_NoDup['Age_of_movie'] > 75) & (IMDB_data_NoDup['Age_of_movie'] <= 100),
     (IMDB_data_NoDup['Age_of_movie'] > 100) & (IMDB_data_NoDup['Age_of_movie'] <= 125),
     (IMDB_data_NoDup['Age_of_movie'] > 125) & (IMDB_data_NoDup['Age_of_movie']<= 150),
     (IMDB_data_NoDup['Age_of_movie'] > 150)]

# create a list of the values we want to assign for each condition
values = ['D0','D1', 'D2', 'D3', 'D4', 'D5', 'D6', 'D7', 'D8']

IMDB_data_NoDup['Decade']  = np.select(conditions, values) 

IMDB_data_NoDup.head()
    

IMDB_data_NoDup.shape

IMDB_data_NoDup['Decade'].value_counts()

IMDB_data_NoDup[IMDB_data_NoDup['Decade'] == '0']

#IMDB_data_NoDup['Genre','TitleType2'] = IMDB_data_NoDup.genres.str.split(",", expand=True)
#df[['First','Last']] = df.Name.str.split("_",expand=True)
#IMDB_data_NoDup['First','Last'] = IMDB_data_NoDup.genres.apply(lambda x: pd.Series(str(x).split(",")))
#IMDB_data_NoDup.head()

# Checking the info of the dataframe


IMDB_data_NoDup.info()

print(IMDB_data_NoDup.isnull().values.sum())

print(IMDB_data_NoDup.isnull().sum())

 #IMDB_data_NoDup.fillna(IMDB_data_NoDup['primaryProfession_director'].value_counts().index[0])

print(IMDB_data_NoDup.isnull().values.sum())

# selecting the values by removing the duplicate value and selecting unique values on the basisi of 
#title", "startYear", "primaryName_Director", "birthYear_Director"

DF_uniquevalue = IMDB_data_NoDup.value_counts(["title", "startYear", "primaryName_Director", "birthYear_Director" ]) 
DF_uniquevalue                                                        #selection of dropping column

# Dropping the duplicate cols after selecting the uniques values from each cols

IMDB_data_NoDup = IMDB_data_NoDup.drop_duplicates(subset=["title", "startYear", "primaryName_Director", "birthYear_Director"])
IMDB_data_NoDup.head()

IMDB_data_NoDup.shape

# verification of the duplicationg of the cols.

Ver_dup = IMDB_data_NoDup[IMDB_data_NoDup['title'] == "Pilot"]
Ver_dup

IMDB_data_NoDup['title'].value_counts().count()

IMDB_data_NoDup.info()

print(IMDB_data_NoDup.isnull().sum())

b = IMDB_data_NoDup[IMDB_data_NoDup['title'] == "Pilot"]                       # Checking pilot(duplicate)
b.shape

IMDB_data_NoDup_copy = IMDB_data_NoDup.copy()

IMDB_data_NoDup_copy.info()

# We can see the missing values in birthYear_Director, deathYear and Director_age. In next step we will 
#replace all missing values with mean values of the respective columns

# checking the value and percentage of missing values

#IMDB_data_NoDup_copy[IMDB_data_NoDup_copy['startYear'].isnull().sum()
IMDB_data_NoDup_copy.isnull().sum()

IMDB_data_NoDup_copy.isna().sum()/(len(IMDB_data_NoDup_copy))*100

IMDB_data_NoDup_copy[IMDB_data_NoDup_copy['startYear'].isnull()].head(2)        # priting Coluumn with null values

IMDB_data_NoDup_copy[IMDB_data_NoDup_copy.isnull().any(axis=1)].head(2)      #Checking Null values in df

 # replacing startyear Null value with mean
IMDB_data_NoDup_copy['startYear'].fillna((IMDB_data_NoDup_copy['startYear'].mean()), inplace=True)

# verifying mean Replacement

IMDB_data_NoDup_copy[IMDB_data_NoDup_copy['title'] == "The Wonderful World of Wilbur Pope"]    
IMDB_data_NoDup_copy[IMDB_data_NoDup_copy['title'] == "Barnyard WAAC"]  

# Checkin Null values in Column birthYear_Director 

IMDB_data_NoDup_copy[IMDB_data_NoDup_copy['birthYear_Director'].isnull()].head(2)   

#Replacing Null values with mean
#IMDB_data_NoDup['startYear'].fillna((IMDB_data_NoDup['startYear'].mean()), inplace=True)

IMDB_data_NoDup_copy['birthYear_Director'] = IMDB_data_NoDup_copy['birthYear_Director'].astype('float')
IMDB_data_NoDup_copy['deathYear'] = IMDB_data_NoDup_copy['deathYear'].astype('float')
IMDB_data_NoDup_copy['Director_age'] = IMDB_data_NoDup_copy['Director_age'].astype('float')



IMDB_data_NoDup_copy['birthYear_Director'].fillna((IMDB_data_NoDup_copy['birthYear_Director'].mean()), inplace = True)
IMDB_data_NoDup_copy['deathYear'].fillna((IMDB_data_NoDup_copy['deathYear'].mean()), inplace = True)
IMDB_data_NoDup_copy['Director_age'].fillna((IMDB_data_NoDup_copy['Director_age'].mean()), inplace = True) 
 

# checking columns after fillna 
IMDB_data_NoDup_copy[IMDB_data_NoDup_copy['primaryName_Director'] == "Francisco Pinto Moreira"] 

# Checking NUll values after replacemnet
    #IMDB_data_NoDup_copy[IMDB_data_NoDup_copy['startYear'].isnull()]
IMDB_data_NoDup_copy[IMDB_data_NoDup_copy.isnull().any(axis=1)].head(2)          


# checking value after replacing all null values
IMDB_data_NoDup_copy.info()                        

# value count for duplicates 
IMDB_data_NoDup_copy['title'].value_counts() 

IMDB_data_NoDup_copy.to_csv('IMDB_data_NoDup_copy_for_prepossing.csv')
